+++ src/FluxFlow.Engine/Core/IFlowEventSource.cs
@@ -1,8 +1,0 @@
-using System.Threading.Tasks.Dataflow;
-
-namespace FluxFlow.Engine.Components;
-
-public interface IFlowEventSource
-{
-    ISourceBlock<FlowEvent> Events { get; }
-}