+++ src/FluxFlow.Engine/Definitions/WellKnownScopes.cs
@@ -1,6 +1,0 @@
-namespace FluxFlow.Engine.Definitions;
-
-public static class WellKnownScopes
-{
-    public const string Resources = "$resources";
-}