+++ src/FluxFlow.Engine/Hosting/FlowApplicationHostBuildErrorCode.cs
@@ -1,7 +1,0 @@
-namespace FluxFlow.Engine;
-
-public enum FlowApplicationHostBuildErrorCode
-{
-    InvalidConfiguration = 1,
-    StartFailed = 2
-}