+++ src/FluxFlow.Engine/Runtime/IFlowNodeModule.cs
@@ -1,6 +1,0 @@
-namespace FluxFlow.Engine.Runtime;
-
-public interface IFlowNodeModule
-{
-    IReadOnlyCollection<IFlowNodeRegistration> Registrations { get; }
-}