+++ src/FluxFlow.Engine/Runtime/RuntimeNodeFactory.cs
@@ -1,5 +1,0 @@
-using FluxFlow.Engine.Definitions;
-
-namespace FluxFlow.Engine.Runtime;
-
-public delegate RuntimeNode RuntimeNodeFactory(RuntimeNodeFactoryContext context);